Extend operation with target to TargetProcessor.
